Now I've implemented compute shader code that will generate a map correspondent to the actual tilemap when the map is generated, upon clicking play  the CPU will pass the livingcell dictionary to the compute shader as a int2 read only structured buffer. The compute shader will then wipe its saved board state, then update the living cells in its 1 dimensional RWStructuredBuffer containing data on all cells.
There is also code being successfully implemented that asynchronously sends the computeshader's grid data back to a readable array of uints  on the CPU side.

When the user plays, the current board state is sent to the GPU. Later code will run GOL on the gpu using this data and send that image directly to a texture instead of running on a tilemap.
When the cmr hits pause (to edit the map) the data is sent back from the GPU to the CPU, the CPU's livingCells data is then updated, and the cmr can edit the board (via a tilemap) as normal.