Updated Compute shader to do the following:
On game start, generate map in compute shader, initialize all values to 0.

When player clicks play for the first time, cpu sends compute shader all living cells. The computeshader stores a snapshot of the board state in AllCellsBufferCPUSnapshot.
on every cpu tick of UpdateCellLifeCycle, compute shader runs GOL in the background on another buffer called AllCellsBufferCurrent
When player clicks Pause, compute shader compares AllCellsBufferCurrent to AllCellsBufferCPUSnapshot, and sends a buffer of cell data back to the CPU that only contains data on the cells that changed between the CPUSnapshot and the current board state.
the CPU then sends newly living cells to livingCells dict, and updates allCells dict accordingly.
Now user can update  boardstate via the already implemented cpu code, and when they click play a new CPUSnapshot will be sent to the compute shader to run GOL on itself