Changed livingcells implementation to instead send the GPU ALL cells.
This build is ready for release, but some future steps:

Create a loading bar and splash screen for when app is loading.

Move all Edit Mode interaction and logic (wherein it changes the buffers and render texture) to the GPU compute shader rather than the CPU, this will make the whole app run as smoothly as the actual GOL implementation currently runs.